refactor: make some changes
- directory structure
- describe the type of the return value in `useCounter.tsx`
- move render to each test case from `beforeEach` in `Counter.test.tsx`
- remove duplicated code in `Counter.test.tsx`
- comments (more readable)
